requirements.txt
aiohttp==3.6.2
aiosqlite==0.15.0
bcrypt==3.2.0
fastapi==0.61.1
pydantic==1.6.1
passlib==1.7.4
PyJWT==1.7.1
python-jose==3.2.0
python-multipart==0.0.5
tortoise-orm==0.16.14
uvicorn==0.11.8
uvloop==0.14.0
